requirements.txt
aniso8601==9.0.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
pika==1.2.0
Pillow==8.3.2
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.7
Werkzeug==2.0.2
